+++ src/components/burger-ingredients-group/burger-ingredients-group.module.css
@@ -1,0 +1,4 @@
+.ingredientsGrid {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+}
