+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -1,8 +1,3 @@
-.ingredientsGrid {
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-}
-
 .container {
     display: flex;
     flex-direction: column;
